workspaceThumbnails: allow requesting size at any time
The slide of thumbnailWorkspace is shown when entering overview, connecting to
the same signal that creates the thumbnails, the showing signal of overview,
but, to make the slide animation we need to know how much width the slider has.
To do that we ask the thumbnailsWorkspace about its width, but given that it
connects to the same signal it could ask the width withouth having created the
thumbnails yet, so reporting a width of 0 and confusing the slide animation.

Currently it acccidentaly works because gjs calls the callbacks following the
order of the clients connecting that signal, and accidentally the
thumbnailsWorskpace is connected before the slide ones.

To avoid that we allow to request the preferred size of the thumbnailsBox at any
time with any number of thumbnails. The only thing required is to make sure the
porthole is accesible when requesting the preferred size.